Enum terminal_cli::AutocompleteResult

pub enum AutocompleteResult {
    None,
    SingleMatch {
        line: AutocompleteLine,
    },
    MultipleMatches {
        lines: Vec<AutocompleteLine>,
    },
}

Result of the autocomplete request on a given set of commands

Variants

None

No suggestions available

SingleMatch

A single match has been found, the line buffer can be immediately expanded with the new command

Fields

line

MultipleMatches

Multiple matches, usually they can be presented to the end user in a column format.

Fields

lines